backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinical-voice-ai-backend"
version = "0.1.0"
description = "FastAPI backend for a voice-driven clinical AI platform."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "psycopg[binary]>=3.2.3,<4.0.0",
  "alembic>=1.14.0,<2.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "bcrypt==4.0.1",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "pydantic-settings>=2.6.1,<3.0.0",
  "python-multipart>=0.0.18,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "sarvamai>=0.1.27,<1.0.0",
  "openai>=1.57.0,<2.0.0",
  "ollama>=0.4.4,<1.0.0",
  "faster-whisper>=1.1.0,<2.0.0",
  "faiss-cpu>=1.9.0,<2.0.0",
  "sentence-transformers>=3.3.1,<4.0.0",
  "numpy>=2.1.3,<3.0.0",
  "reportlab>=4.4.10,<5.0.0",
  "tenacity>=9.0.0,<10.0.0",
  "structlog>=24.4.0,<25.0.0",
  "orjson>=3.10.12,<4.0.0"
]

[tool.setuptools.packages.find]
include = ["app*"]
